:recycle: [refactor] Controller 도입 및 Model과 View의 책임 범위 수정
